% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.model.R
\name{bias_correct}
\alias{bias_correct}
\title{Applies bias correction with a specified method to the data and saves or returns scores.}
\usage{
bias_correct(dt, method, par_1, ...)
}
\arguments{
\item{dt}{The data table.}

\item{method}{Method of bias correction. Takes "sma" for simple moving average and "ema" for exponential moving average.}

\item{par_1}{Numeric. If method == "sma", par_1 is the (integer) length of the moving average window, if method == "ema", par_1 is the scale parameter for the exponential downscaling, typically in (0,1).}

\item{...}{arguments passed on to the moving average function, e.g. two_sided for the training period.}
}
\value{
A data table containing columns .(year,month,Lon,Lat,Bias_est,SST_hat).
}
\description{
Applies bias correction with a specified method to the data and saves or returns scores.
}
\author{
Claudio Heinrich
}
