% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.model.R
\name{bias_correct_2}
\alias{bias_correct_2}
\title{Applies bias correction with a specified method to the data and saves or returns scores.}
\usage{
bias_correct_2(dt, method, par_1, skip = 0, reduced_output = TRUE)
}
\arguments{
\item{dt}{The data table.}

\item{method}{Method of bias correction. Takes "sma" for simple moving average and "ema" for exponential moving average.}

\item{par_1}{Numeric. If method == "sma", par_1 is the (integer) length of the moving average window, if method == "ema", par_1 is the scale parameter for the exponential downscaling, typically in (0,1).}

\item{skip}{Integer. Passed on to sim_mov_av or exp_mov_av.}

\item{scores}{Logical. If true, the MSE is returned.}

\item{eval_years}{Numerical vector. The years for evaluating the score.}

\item{saveorgo}{Logical. If TRUE, the data table with corrected SST_hat and new column Bias_Est is saved.}

\item{save_dir, file_name}{Directory and name for the saved file.}
}
\value{
The data table with corrected SST_hat and new column Bias_Est.
}
\description{
Applies bias correction with a specified method to the data and saves or returns scores.
}
\examples{
\dontrun{bias_correct(saveorgo = FALSE)}


}
\author{
Claudio Heinrich
}
