% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.model.R
\name{bias_lr_bb_par}
\alias{bias_lr_bb_par}
\title{parallelized version of bias_lr_bb}
\usage{
bias_lr_bb_par(DT, months = 1:12, validation_years = 2001:2016,
  mc_cores)
}
\arguments{
\item{DT}{the data table.}

\item{months}{The considered months.}

\item{validation_years}{Validation years.}

\item{...}{Arguments passed on to mclapply, most importantly mc.cores.}
}
\description{
parallelized version of bias_lr_bb
}
\examples{
\dontrun{ DT = load_combined_wide()
                    bias_lr_bm(DT = DT)}
  
}
\author{
Claudio Heinrich
}
