% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.model.R
\name{bias_lr_bm}
\alias{bias_lr_bm}
\title{computes the RMSE when instead of bias correction the linear model SST_hat = a + b Ens_bar is used (as in standard NGR), with data grouped by month.}
\usage{
bias_lr_bm(DT, months = 1:12, validation_years = 2001:2010)
}
\arguments{
\item{DT}{the data table.}

\item{months}{The considered months.}

\item{training_years, validation_years}{Training years and validation years.}
}
\description{
computes the RMSE when instead of bias correction the linear model SST_hat = a + b Ens_bar is used (as in standard NGR), with data grouped by month.
}
\examples{
\dontrun{ DT = load_combined_wide()
                    bias_lr_bm(DT = DT)}
  
}
\author{
Claudio Heinrich
}
