% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.model.R
\name{bias_lr_bm_la}
\alias{bias_lr_bm_la}
\title{computes the MSE when instead of bias correction the locally adaptive model SST_ano = a + b Ens_bar_ano is used (as in standard NGR), with data grouped by month.}
\usage{
bias_lr_bm_la(DT, months = 1:12, validation_years = 2001:2010)
}
\arguments{
\item{DT}{the data table.}

\item{months}{The considered months.}

\item{validation_years}{validation years.}
}
\description{
computes the MSE when instead of bias correction the locally adaptive model SST_ano = a + b Ens_bar_ano is used (as in standard NGR), with data grouped by month.
}
\examples{
\dontrun{ DT = load_combined_wide()
                    bias_lr_bm(DT = DT)}
  
}
\author{
Claudio Heinrich
}
