% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.functions.R
\name{bootstrap_difference}
\alias{bootstrap_difference}
\title{Run a bootstrap test of the pairwise difference between two vectors of numbers}
\usage{
bootstrap_difference(sc1, sc2, q_prob, N = 1000)
}
\arguments{
\item{sc1}{Vector, the scores from one method}

\item{sc2}{Vector, the scores from some other method}

\item{q_prob}{a vector containing probabilities for which the quantiles should be returned from the bootstrap distribution}

\item{N}{Integer, the size of the bootstrap distribution}
}
\value{
A list as returned by the function boot::boot, with additional values p_val and q
}
\description{
Run a bootstrap test of the pairwise difference between two vectors of numbers
}
\examples{
N = 1e2
trend  = 1:N
sc1 = trend + .01 + rnorm(N, .001)
sc2 = trend - .01 + rnorm(N, .001)
l = bootstrap_difference(sc1,sc2)
}
\author{
Claudio
}
