% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{combine_data_wide}
\alias{combine_data_wide}
\title{Align and combine an ensemble (NorCPM) and an observation (SeNorge) set of data and store this in a wide format}
\usage{
combine_data_wide(dt_ens, dt_obs, dt_map)
}
\arguments{
\item{dt_ens}{The data.table of ensemble memebers}

\item{dt_obs}{The data.table of observational members}

\item{dt_map}{A data.table that maps the two between one another}
}
\description{
This aligns a collection of ensemble members with the associated observational member to create a single dataset.  Data are stored in "wide" format, indicating that all 9 ensemble members and all 10 observational outcomes for a given y,m,grid location are stored in one line of data with a whole bunch of columns.
}
\examples{
##dt_ens = load_ensemble(1990,1)
##dt_obs = load_observations(1990,1)
##dt_map = load_mapping()
##combine_data_wide(dt_ens,dt_obs,dt_map)
}
\author{
Alex Lenkoski
}
