% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.model.R
\name{compute_clim}
\alias{compute_clim}
\title{computes climatology and sample climatology, always oos starting from the last year}
\usage{
compute_clim(DT)
}
\arguments{
\item{DT}{the data table.}
}
\description{
computes climatology and sample climatology, always oos starting from the last year
}
\author{
Claudio Heinrich
}
