% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{contruct_grid_map}
\alias{contruct_grid_map}
\title{Create a mapping between the grid points on the
NorCPM ensemble and the SeNorge observational grid}
\usage{
contruct_grid_map(dt_ens = load_ensemble(1985, 1),
  dt_obs = load_observations(1985, 1))
}
\arguments{
\item{dt_ens}{A data.table of the ensemble members}

\item{dt_obs}{A data.table of the observation members}
}
\value{
A datatable which maps each lat/lon in the ensemble space to the nearest
member in the observation space
}
\description{
This helper function performs a nearest neighbor analysis to find the grid point in the NorCPM ensemble
that is closest to each point in the SeNorge dataset.
}
\examples{
##construct_grid_map()
}
\author{
Alex Lenkoski
}
