% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE.R
\name{cov_est_SE}
\alias{cov_est_SE}
\title{computes the covariance estimate by the eigenvalue-shrinking-method of Ledoit and Wolf, 
but with the sample covariance matrix downweighted by distance of gridpoints before deriving the estimate. For some details and notation see Pourahmadi 2013, pp.101-102.}
\usage{
cov_est_SE(dt, weight_mat, Y = unique(dt[, year]), M = 1:12,
  save_years = 2013:2018, save_dir, file_name = "cov_est_SE")
}
\arguments{
\item{dt}{The data table}

\item{weight_mat}{The matrix containing the weights for tapering: Has dimension n x n, where n is the number of grid_ids in water in dt. 
Should be generated by the function weight_mat.}

\item{Y, M}{The years and months considered.}

\item{save_years}{For these years the covariance estimate is computed and saved,
using for each year all previous years.}

\item{save_dir}{The directory to save in.}

\item{file_name}{The saved files are named paste0(file_name,"_m",+,"_y",++,".RData"), where + is the month and ++ the year.}
}
\description{
computes the covariance estimate by the eigenvalue-shrinking-method of Ledoit and Wolf, 
but with the sample covariance matrix downweighted by distance of gridpoints before deriving the estimate. For some details and notation see Pourahmadi 2013, pp.101-102.
}
