% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.functions.R, R/univariate.model.R
\name{crps_na_rm}
\alias{crps_na_rm}
\title{crps score for normal distribution, resistant to missing values and 0 standard deviation}
\usage{
crps_na_rm(y, mean, sd)

crps_na_rm(y, mean, sd)
}
\arguments{
\item{y}{vector of observations.}

\item{mean, sd}{mean and sd of the forecasting distribution}

\item{y}{vector of observations.}

\item{mean, sd}{mean and sd of the forecasting distribution}
}
\value{
vector of the same length as y containing crps scores.

vector of the same length as y containing crps scores.
}
\description{
crps score for normal distribution, resistant to missing values and 0 standard deviation

crps score for normal distribution, resistant to missing values and 0 standard deviation
}
\examples{
crps_na_rm(c(NA,rnorm(10)), 1,1)

crps_na_rm(c(NA,rnorm(10)), 1,1)

}
\author{
Claudio Heinrich

Claudio Heinrich
}
