% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_dt.R
\name{dt_transform_center}
\alias{dt_transform_center}
\title{Transforming the datatable (which has been univariately post-processed) by replacing temperatures by temperature anamolies}
\usage{
dt_transform_center(dt, clim_years, trafo_colnames = c("SST_bar", "SST_hat",
  "Ens_bar", paste0("Ens", 1:9)))
}
\arguments{
\item{dt}{the data table.}

\item{clim_years}{The years of climatology.}
}
\value{
the transformed data table dt, containing an extra column \code{clim}.
}
\description{
replaces all temperature-valued columns (forecasts and observations) in dt by anomalies w.r.t. the mean observed temp.
             over all climatology years. For the climatology years it leaves the current year out (out of sample).
}
\examples{
\dontrun{}

}
\author{
Claudio Heinrich
}
