% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_dt.R
\name{dt_transform_stan}
\alias{dt_transform_stan}
\title{Transforming the datatable (which has been univariately post-processed) by replacing temperatures by standardized temperature anamolies}
\usage{
dt_transform_stan(dt, clim_years, crit_ts = 0.01)
}
\arguments{
\item{dt}{the data table.}

\item{clim_years}{The years of climatology.}

\item{crit_ts}{Locations with climatology standard deviation below this threshold are transformed into NAs}
}
\value{
the transformed data table dt, containing extra columns \code{clim} and \code{clim_sd}.
}
\description{
replaces all temperature-valued columns (forecasts and observations) in dt by standardized anomalies w.r.t. 
             the mean and sd of theobserved temp. over all climatology years. For the climatology years it leaves the current
             year out (out of sample). Locations with (almost) 0 sd in climatology get assigned NAs.
}
\examples{
\dontrun{}

}
\author{
Claudio Heinrich
}
