% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.model.R
\name{ens_sd_est}
\alias{ens_sd_est}
\title{sets up the estimation of marginal variance by computing the ensemble variance by location, month and year.}
\usage{
ens_sd_est(dt, ens_size = 9, saveorgo = TRUE, mean_est = "sv",
  save_dir = "~/PostClimDataNoBackup/SFE/Derived/",
  file_name = "dt_combine_wide_bias_sd.RData")
}
\arguments{
\item{dt}{The data table containing the estimated bias.}

\item{ens_size}{integer. Size of the ensemble.}

\item{saveorgo}{logical. Do we save the data table?}

\item{mean_est}{}

\item{save_dir, file_name}{character strings. Directory and file name for saving.}
}
\value{
the data table dt with a new column labelled 'var_bar'
}
\description{
sets up the estimation of marginal variance by computing the ensemble variance by location, month and year.
}
\examples{
\dontrun{load_combined_wide()
         ens_sd_est(dt)}
         
}
\author{
Claudio Heinrich
}
