% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.model.R
\name{exp_mov_av}
\alias{exp_mov_av}
\title{computes exponentially weighted moving averages, is resistant to missing values}
\usage{
exp_mov_av(a, vec, years, skip = 0, twosided = FALSE)
}
\arguments{
\item{a}{weight parameter.}

\item{vec, years}{vectors of the same length, vec[i] contains the value corresponding to year years[i]}
}
\value{
a vector of the same length as the input vectors. At location j it contains the average of the past entries of vec, weighted by exp(-a*d) where d is the distance to the current year. First entry is 0.
}
\description{
computes exponentially weighted moving averages, is resistant to missing values
}
\examples{
exp_mov_av(.1, rnorm(10), c(1990,1993,1995:2002))


}
\author{
Claudio Heinrich
}
