% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECC.R
\name{forecast_ECC}
\alias{forecast_ECC}
\title{ECC forecasts}
\usage{
forecast_ECC(dt, Y = NULL, M = NULL, ens_size = 9, method = "Q")
}
\arguments{
\item{dt}{the data table containing bias and variance estimates.}

\item{Y, M}{vectors contining the years and months for the forecast, default are all years and months in dt.}

\item{ens_size}{Integer. Size of the NWP ensemble.}

\item{method}{takes "R" or "Q". Method of ECC, see Schefzik et al. 2013.}
}
\value{
data table containing \code{ens_size} forecast columns, labelled "fc"1:ens_size, and the corresponding noise columns, defined as fc - SST_hat
}
\description{
Generates forecasts post-processed by ensemble copula coupling (ECC)
}
\examples{
\dontrun{}

}
\author{
Claudio Heinrich
}
