% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostationary.R
\name{forecast_GS}
\alias{forecast_GS}
\title{geostationary forecasts}
\usage{
forecast_GS(dt, Y, M, n = 10, mc_cores = 12, noise = FALSE, var_dir,
  var_file_name = "variogram_exp")
}
\arguments{
\item{dt}{the data table.}

\item{Y, M}{Integer vectors containing year(s) and month(s).}

\item{n}{Integer. Size of the desired forecast ensemble.}

\item{noise}{Logical. Should the return data contain columns with noise?}

\item{var_dir, var_file_names}{Where the fitted variograms are stored.}
}
\value{
data table containing n columns with forecasts.
}
\description{
Generates forecasts using a geostationary model with exponential covariance function for post-processing
}
\author{
Claudio Heinrich
}
