% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{forecast_PCA_mult_corr}
\alias{forecast_PCA_mult_corr}
\title{Forecasts by PCA}
\usage{
forecast_PCA_mult_corr(dt, Y, M, n = 10, nPCs, mc_cores = length(M),
  noise = FALSE, cov_dir, cov_file_name = "PCA_cov")
}
\arguments{
\item{dt}{the data table.}

\item{Y, M}{Integer vectors containing year(s) and month(s).}

\item{n}{Integer. Size of the desired forecast ensemble.}

\item{nPCs}{Integer vector containing the numbers of considered principal components.}

\item{cov_dir, cov_file_name}{Where the PCA covariance matrices are stored and how they are named, see \code{sam_cov}.}
}
\value{
data table containing n columns with noise and n columns with forecasts.
}
\description{
Generates forecasts for the PCA post-processing approach.
}
\examples{
\dontrun{}

}
\author{
Claudio Heinrich
}
