% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schaake.R
\name{forecast_Schaake}
\alias{forecast_Schaake}
\title{Schaake shuffle forecasts}
\usage{
forecast_Schaake(dt, obs_ens_years, Y = NULL, M = NULL, method = "Q")
}
\arguments{
\item{dt}{the data table containing bias and variance estimates.}

\item{obs_ens_years}{years to be used as historical archive of observations to determine dependence structure.}

\item{Y, M}{vectors contining the years and months for the forecast, default are all months in dt, and all years that are not contained in obs_ens_years.}

\item{method}{takes "R" or "Q", similar as ECC, see forecast_ECC.}
}
\value{
data table containing \code{ens_size = length(obs_ens_years)} forecast columns, labelled "fc"1:ens_size, and the corresponding noise columns, defined as fc - SST_hat
}
\description{
Generates forecasts post-processed by Schaake shuffle
}
\author{
Claudio Heinrich
}
