% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE.R
\name{forecast_SE}
\alias{forecast_SE}
\title{Forecasts by shrinking eigenvalues}
\usage{
forecast_SE(dt, Y, M, n = 10, cov_dir, cov_file_name = "cov_est_SE")
}
\arguments{
\item{dt}{the data table.}

\item{Y, M}{Integer vectors containing year(s) and month(s).}

\item{n}{Integer. Size of the desired forecast ensemble.}

\item{cov_dir, cov_file_name}{Where the covariance estimates are stored and how they are named, see \code{cov_est_SE}.}
}
\value{
data table containing n columns with noise and n columns with forecasts.
}
\description{
Generates forecasts for the SE post-processing approach.
}
\examples{
\dontrun{}

}
\author{
Claudio Heinrich
}
