% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostationary.R
\name{geostationary_training}
\alias{geostationary_training}
\title{preparation for geostationary forecasts}
\usage{
geostationary_training(dt, training_years = 1985:2000, m = 1:12, save_dir,
  file_name = "variogram_exp_m", nintv = 75, truncate = TRUE,
  mc_cores = 1)
}
\arguments{
\item{dt}{the data table.}

\item{training_years, m}{Integer vectors containing the year(s) and month(s) of the training dataset.}

\item{save_dir, file_name}{The directory to save in. The name of the file of the saved variogram for a given month is \code{file_name <month> .RData}.}

\item{nintv}{Integer. How many distance bins are considered for the empirical variogram.}

\item{truncate}{Logical. The empirical variogram oftentimes is far from the fitted variogram for the 10% largest distances considered. If truncate == TRUE, those are ignored leading to a visually much better fit of the variogram.}

\item{ens_size}{Integer. Size of the NWP ensemble.}

\item{saveorgo}{Logical, whether we save or not.}
}
\value{
data table containing n columns with noise and n columns with forecasts.
}
\description{
Estimates the variogram for the residuals assuming anexponential covariance model and saves the results.
}
\examples{
\dontrun{ DT = load_combined_wide(data_dir = "~/PostClimDataNoBackup/SFE/Derived/NAO", output_name = "dt_combine_NAO_wide_bc.RData") \\cr
                    geostationary_training(dt = DT)}

}
\author{
Claudio Heinrich, Yuan Qifen
}
