% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.model.R
\name{global_mean_scores_EnsBMA}
\alias{global_mean_scores_EnsBMA}
\title{Computes scores for bias correction and variance estimation}
\usage{
global_mean_scores_EnsBMA(DT, ens_size = 9, eval_years = 2001:2010,
  var = TRUE)
}
\arguments{
\item{DT}{The data table}

\item{ens_size}{Size of the ensemble.}

\item{eval_years}{The years to compute the scores for}

\item{var}{Logical. If TRUE the CRPS is computed, else the MSE of the bias corrected forecast is computed}
}
\value{
A one-row data table containing the mean score.
}
\description{
Computes scores for bias correction and variance estimation
}
\examples{
\dontrun{
save_dir = "~/PostClimDataNoBackup/SFE/Derived/NAO/"
DT = load_combined_wide(data_dir = save_dir, output_name = "dt_combine_wide_bc_var.RData")
global_mean_scores(DT)}


}
\author{
Claudio Heinrich
}
