% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiv.validation.R
\name{hist_transform}
\alias{hist_transform}
\title{histogram transformation}
\usage{
hist_transform(vec)
}
\arguments{
\item{vec}{A vector containing ranks.}
}
\value{
A transformed vector containing data on [0,1]
}
\description{
when working with rank histograms problems arise when the number of bins does not divide the number of possible ranks
This function solves this problem by randomizing the data such that it is uniform on [0,1] if it was uniform before
}
\author{
Claudio Heinrich
}
