% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_combined_wide}
\alias{load_combined_wide}
\title{Load a pre-rendered combined wide dataset}
\usage{
load_combined_wide(data_dir = "~/PostClimDataNoBackup/SFE/Derived/",
  vintage = "mr", bias = FALSE, var = FALSE, output_name = NULL)
}
\arguments{
\item{data_dir}{Root location of the data}

\item{vintage}{Indication of which vintage we are using, most likely 'mr'}

\item{bias}{Boolean. Should the bias corrected ensemble be loaded?}

\item{var}{Boolean. Should the data with estimated bias and variance be loaded?}

\item{model}{String.  Currently taking values \code{NorESM} or \code{senorge} to indicate whether the NorESM or senorge data are of interest}
}
\value{
A data.table containing all data for the parameters specified, provided these data have been rendered.
}
\description{
Given a pre-rendered dataset, this loads it from the correct location
}
