% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_ensemble}
\alias{load_ensemble}
\title{Load and postprocess a given NorCPM ensemble member}
\usage{
load_ensemble(year, month, vintage = "mr",
  data_dir = "~/PostClimDataNoBackup/")
}
\arguments{
\item{year}{Integer.  The year you are interested in.}

\item{month}{Integer.  The month}

\item{vintage}{One of Jan, Apr, Jul, Oct or mr.  If mr, the vintage closest to the given month is used.}

\item{data_dir}{String.  The root directory that stores the data.  Data are then assumed stored in \code{SFE/NorCPM_Ocean/}}
}
\description{
This loads a given NorCPM ensemble member given the correct directory structure
}
\examples{
##load_ensemble(1990,1)

}
