% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_observations}
\alias{load_observations}
\title{Load and postprocess a given SeNorge observational member}
\usage{
load_observations(year, month, obsdir = "./Data/HadiSST2/")
}
\arguments{
\item{year}{Integer.  The year you are interested in.}

\item{month}{Integer.  The month}

\item{obsdir}{String.  The root directory that stores the data.}
}
\description{
This loads a given SeNorge observational member given the correct directory structure
}
\examples{
##load_observations(1990,1)

}
