% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/production.R
\name{make_combined_wide_dataset}
\alias{make_combined_wide_dataset}
\title{Make the combined dataset for all years and store in a wide format}
\usage{
make_combined_wide_dataset(y_start = 1985, y_stop = 2010, vintage = "mr",
  data_dir = "~/PostClimDataNoBackup/",
  grid_mapping_loc = "~/PostClimDataNoBackup/SFE/Derived/",
  output_loc = "~/PostClimDataNoBackup/SFE/Derived/", output_name = NULL,
  lat_box = c(-Inf, Inf), lon_box = c(-Inf, Inf))
}
\arguments{
\item{y_start}{The first year for which we have obs/ens data}

\item{y_stop}{The final year for which we have obs/ens data}

\item{vintage}{Which vintage are we using, often "mr" for most recent}

\item{data_dir}{The root directory that stores our data}

\item{grid_mapping_loc}{The location where the precomputed grid alignment object is stored.}

\item{output_loc}{The folder where the output should be stored. If \code{NULL} then put in the same location as \code{data_dir}}

\item{output_name}{The name of the output file.  If \code{NULL} then \code{paste0("dt_combine_",vintage,"_wide.RData")} will be used}

\item{lat_box}{The upper and lower bounds for the latitude.  By default the entire globe is used.}

\item{lon_box}{The upper and lower bounds for the longitude.  By default the entire globe is used.}
}
\description{
Make the combined dataset for all years and store in a wide format
}
