% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/production.R
\name{make_grid_mapping}
\alias{make_grid_mapping}
\title{A utility function that makes and stores the grid mapping}
\usage{
make_grid_mapping(out.file = "./Data/PostClim/SFE/Derived")
}
\arguments{
\item{out.file}{Location to store the output, which is then saved at \code{/dt_map.RData}}
}
\description{
This is essentially a wrapper for \code{construct_grid_map}
}
