% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiv.validation.R
\name{mv_rank_hist}
\alias{mv_rank_hist}
\title{making and plotting of multivariate rank histograms}
\usage{
mv_rank_hist(dt_fc, fc_ens_size, mn = "", breaks = 10,
  save_pdf = FALSE, plot_dir = "", file_name = "")
}
\arguments{
\item{dt_fc}{the forecast data table.}

\item{fc_ens_size}{Size of the forecast ensemble: dt_fc ought to have columns named fc1,...,fcn, where n = fc_ens_size.}

\item{mn}{Title of the plot}

\item{breaks}{Number of breaks in the rank histograms. Should at most be 1 + fc_ens_size (which is the size of the ensemble for ECC).}

\item{save_pdf, plot_dir, file_name}{Whether and where the rank histogram should be saved as pdf. \code{file_name} should not contain '.pdf'.}
}
\value{
none, if save_pdf = FALSE, the rank histogram is shown
}
\description{
computes the multivariate rank histograms for the provided forecast data table. Plots the average rank, the band depth rank, and the minimum spanning tree rank.
}
\examples{
\dontrun{}

}
\author{
Claudio Heinrich
}
