% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA_cov}
\alias{PCA_cov}
\title{computes and saves the SVD of the sample covariance matrices downweighted by distance of gridpoints}
\usage{
PCA_cov(dt, weight_mat, Y = unique(dt[, year]), M = 1:12, nPCs,
  save_years = 2013:2018, save_dir, file_name = "PCA_cov")
}
\arguments{
\item{dt}{The data table}

\item{weight_mat}{The matrix containing the weights for tapering: Has dimension n x n, where n is the number of grid_ids in water in dt. 
Should be generated by the function weight_mat.}

\item{Y, M}{The years and months considered.}

\item{nPCs}{the max number of principal components to consider.}

\item{save_years}{For these years sam_cov generates and saves the sample covariance matrices, 
using for each year all previous years.}

\item{save_dir}{The directory to save in.}

\item{file_name}{The saved files are named paste0(file_name,"_m",+,"_y",++,".RData"), where + is the month and ++ the year.}
}
\description{
computes and saves the SVD of the sample covariance matrices downweighted by distance of gridpoints
}
