% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.functions.R
\name{permutation_test_difference}
\alias{permutation_test_difference}
\title{Run a permutation test of the pairwise difference between two vectors of numbers}
\usage{
permutation_test_difference(a, b, N = 5000)
}
\arguments{
\item{a}{Vector, the scores from one method}

\item{b}{Vector, the scores from some other method}

\item{N}{Integer, the size of the permutation distribution}
}
\value{
A list with the mean of the difference and the permutation distribution of that difference and the p-value
}
\description{
Run a permutation test of the pairwise difference between two vectors of numbers
}
\examples{
N = 1e2
trend  = 1:N
a = trend + .01 + rnorm(N, .001)
b = trend - .01 + rnorm(N, .001)
l = permutation_test_difference(a,b)
q = sum(l$D <= l$d_bar) / length(l$D)
}
\author{
Alex,Claudio
}
