% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostProcessing-package.R
\docType{package}
\name{pp.sst-package}
\alias{pp.sst}
\alias{pp.sst-package}
\title{pp.sst: postprocessing sea surface temperature forecasts}
\description{
Reproduces the results presented in Heinrich et al. 2019.
}
\author{
\strong{Maintainer}: Claudio Heinrich \email{claudio.heinrich@nr.no}

}
\keyword{internal}
