% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiv.validation.R
\name{rhist_dt}
\alias{rhist_dt}
\title{rank histograms for data tables}
\usage{
rhist_dt(B, max_rk, breaks = 11, hist_xlab = NULL, hist_ylab = NULL,
  hist_ylim = NULL, mn = "")
}
\arguments{
\item{B}{A data table, containing the key variable (most commonly YM) as first column and the ranks of the observations as second}

\item{breaks}{Number of breaks desired for the rank histogram.}

\item{hist_xlab, hist_ylab}{Labels for axis.}

\item{hist_ylim}{Range of y-axis.}
}
\description{
rank histograms for data tables
}
\author{
Claudio Heinrich
}
