% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.model.R
\name{sd_est_2}
\alias{sd_est_2}
\title{Estimates standard deviation with a specified method to the data and saves or returns scores.}
\usage{
sd_est_2(dt, method, par_1, saveorgo = TRUE,
  save_dir = "~/PostClimDataNoBackup/SFE/Derived/",
  file_name = "dt_combine_wide_bc_var.RData", skip = 0,
  reduced_output = TRUE)
}
\arguments{
\item{dt}{The data table.}

\item{method}{Method of variance estimation. Takes "sma" for simple moving average and "ema" for exponential moving average.}

\item{par_1}{Numeric. If method == "sma", par_1 is the (integer) length of the moving average window, if method == "ema", par_1 is the scale parameter for the exponential downscaling, typically in (0,1).}

\item{saveorgo}{Logical. If TRUE, the data table with new column SD_hat is saved.}

\item{save_dir, file_name}{Directory and name for the saved file.}

\item{scores}{Logical. If true, the CRPS is returned.}

\item{eval_years}{Numerical vector. The years for evaluating the score.}
}
\value{
The data table dt containing a new column SD_hat.
}
\description{
Estimates standard deviation with a specified method to the data and saves or returns scores.
}
\examples{
\dontrun{sd_est(saveorgo = FALSE)}


}
\author{
Claudio Heinrich
}
