% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.functions.R
\name{trc}
\alias{trc}
\title{truncation function for freezing level of sea water}
\usage{
trc(x, truncation.value = -1.79)
}
\description{
truncation function for freezing level of sea water
}
\author{
Claudio Heinrich
}
