% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.model.R
\name{var_est_NGR_bl}
\alias{var_est_NGR_bl}
\title{Estimates the variance as has been suggested for NGR, grouped by location}
\usage{
var_est_NGR_bl(dt, months = 1:12, validation_years = 2001:2010,
  mc.cores = 1)
}
\arguments{
\item{dt}{the data table.}

\item{months}{The considered months.}

\item{validation_years}{Training years and validation years.}

\item{mc.cores}{Should we parallelize and with how many cores. Large datasets run into memory issues if mc.cores>1.}
}
\value{
data table containing the RMSEs for the model above, where the coefficients are estimated in three different ways: grouped by month, location and by both.
}
\description{
Estimates the variance as has been suggested for NGR, grouped by location
}
\author{
Claudio Heinrich
}
