% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.model.R
\name{var_est_NGR_bm}
\alias{var_est_NGR_bm}
\title{Estimates the variance as has been suggested for NGR, grouped by month}
\usage{
var_est_NGR_bm(dt, months = 1:12, validation_years = 2001:2010)
}
\arguments{
\item{dt}{the data table.}

\item{months}{The considered months.}

\item{validation_years}{Training years and validation years.}
}
\value{
data table containing the RMSEs for the model above, where the coefficients are estimated in three different ways: grouped by month, location and by both.
}
\description{
Estimates the variance as has been suggested for NGR, grouped by month
}
\author{
Claudio Heinrich
}
