% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiv.validation.R
\name{var_sc_emp}
\alias{var_sc_emp}
\title{computing variogram score for a given month and year}
\usage{
var_sc_emp(m, y, dt_fc, n, p)
}
\arguments{
\item{m, y}{Month and year.}

\item{dt_fc}{The forecast data table.}

\item{n}{number of forecasts available for moment estimation.}

\item{p}{Power used in the variogram score.}
}
\value{
a data table with one row and three columns: year, month and variogram score.
}
\description{
computing variogram score for a given month and year
}
\author{
Claudio Heinrich
}
