% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiv.validation.R
\name{var_sc_par}
\alias{var_sc_par}
\title{computing empirical variogram scores}
\usage{
var_sc_par(dt_fc, years, ms, n, p = 0.5, save_dir = NULL,
  file_name = "vs", mc_cores = NULL)
}
\arguments{
\item{dt_fc}{The forecast data table.}

\item{years, ms}{The years and months for which to compute the variogram score.}

\item{n}{number of forecasts available for moment estimation.}

\item{p}{Power used in the variogram score.}

\item{save_dir, file_name}{If we should save and where. \code{file_name} ought not contain '.RData'.}

\item{mc.cores}{How many cores do you want to use?}
}
\value{
a data table with three columns: year, month and vs.
}
\description{
Computes the variogram scores for each month and year for a given forecast data table. Parallelized along months.
}
\author{
Claudio Heinrich
}
