% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{weight_mat}
\alias{weight_mat}
\title{computes the weight matrix for downweighing the sample covariance matrix by distance.}
\usage{
weight_mat(dt, phi = GneitingWeightFct, L = 2500)
}
\arguments{
\item{dt}{The data table.}

\item{phi}{Weight function, takes parameter \code{L}, i.e. phi = phi(d,L) where d is the distance of the grid points.}

\item{L}{Parameter for the weight function.}
}
\description{
computes the weight matrix for downweighing the sample covariance matrix by distance.
}
